refactor: add interfaces file
