+++ fun-chat/src/components/authentication/login.ts
@@ -1,7 +1,7 @@
 import "./login.css";
 import loginFormRenderer from "./authPageRender";
 import { checkInputs } from "./validation/checkInputs";
-import createUser from "../api/createUser";
+import createUser from "../utilities/createUser";
 
 class LoginForm {
   constructor() {
